internals: Theme.jsonToTailwind()
