+++ apps/~internal/lib/index.ts
@@ -1,4 +1,5 @@
 export * as Env from './Env.js'
 export * as PortoConfig from './PortoConfig.js'
 export * as Query from './Query.js'
+export * from './Theme.js'
 export * as UserAgent from './UserAgent.js'
